+++ client/src/pages/NewDiary.tsx
@@ -3,7 +3,7 @@ import Fade from "react-reveal/Fade";
 
 import { PinkPurpleBackground } from "src/components/common/Backgrounds";
 import BackButton from "src/components/common/BackButton";
-import PostButton from "src/components/new-diary/PostButton";
+import PostButton from "src/components/diary/PostButton";
 import PageTitle from "src/components/common/PageTitle";
 import NewDiaryForm from "src/components/new-diary/NewDiaryForm";
